move StateMachine and states to a new file, etc
